pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "wheel", "moviepy"]
build-backend = "setuptools.build_meta"

[project]
name = "tinyworld"
version = "0.1.0"
authors = [
  { name="A. Aurelions", email="aurelions@protonmail.com" }
]
description = "A small library to create colorful, dynamic videos with TTS audio"
readme = "README.md"
license = { text = "MIT License" }
requires-python = ">=3.7"
dependencies = [
  "gTTS",
  "moviepy",
  "Pillow",
]
keywords = ["video", "TTS", "audio", "visualization"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Multimedia :: Video",
]

